% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesEFA-package.R
\docType{package}
\name{BayesEFA-package}
\alias{BayesEFA-package}
\alias{BayesEFA}
\title{The 'BayesEFA' package.}
\description{
Bayesian Exploratory Factor Analysis using Hamiltonian Monte Carlo
via Stan. Posterior draws are post-processed with Varimax rotation and the RSP
alignment algorithm to resolve rotational indeterminacy. Provides Bayesian fit
indices, reliability estimates (Omega), and factor scores with full uncertainty
quantification.
}
\references{
Stan Development Team (2024). RStan: the R interface to Stan. R package. https://mc-stan.org
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RicardoReySaez/BayesEFA}
  \item Report bugs at \url{https://github.com/RicardoReySaez/BayesEFA/issues}
}

}
\author{
\strong{Maintainer}: Ricardo Rey-Sáez \email{ricardoreysaez95@gmail.com} (\href{https://orcid.org/0000-0001-6739-2035}{ORCID})

}
