% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_posterior.R
\name{posterior_summaries}
\alias{posterior_summaries}
\title{Compute Posterior Summaries}
\usage{
posterior_summaries(befa_fit, pars, ...)
}
\arguments{
\item{befa_fit}{A \code{befa} object.}

\item{pars}{Character vector. Parameter names to summarize.}

\item{...}{Additional arguments passed to \code{\link[posterior:draws_summary]{posterior::summarise_draws()}}.}
}
\value{
A data frame of summaries.
}
\description{
A wrapper around \code{\link[posterior:draws_summary]{posterior::summarise_draws()}} to compute summaries directly
from a \code{befa} object.
}
\examples{
\dontrun{
befa_fit <- befa(
  data = HS_data, n_factors = 3, model = "cor",
  iter = 500, chains = 2, seed = 123
)

# Posterior summaries of communalities
posterior_summaries(befa_fit, pars = "h2")
}

}
